integrasi api (1)

+++ app/Controllers/Login.php
@@ -2,7 +2,6 @@
 
 namespace App\Controllers;
 
-use App\Models\ModelUser;
 
 class Login extends BaseController
 {
@@ -12,7 +11,7 @@ class Login extends BaseController
         $data = [
             'judul' => 'Login'
         ];
-        
+
         echo view('templates/header', $data);
         echo view('auth/login');
         echo view('templates/footer');
@@ -21,30 +20,36 @@ class Login extends BaseController
     public function auth()
     {
         $session = session();
-        $model = new ModelUser();
         $email = $this->request->getVar('email');
         $password = $this->request->getVar('password');
-        $data = $model->where('user_email', $email)->first();
 
-        if($data){
-            $pass = $data['user_password'];
-            $verify_pass = password_verify($password, $pass);
-
-            if($verify_pass){
-                $ses_data = [
-                    'user_id' => $data['user_id'],
-                    'user_name' => $data['user_name'],
-                    'user_email' => $data['user_email'],
-                    'logged_in' => true
-                ];
-                $session->set($ses_data);
-                return redirect()->to(base_url('/'));
-            } else {
-                $session->setFlashdata('msg', 'Wrong Password');
-                return redirect()->to(base_url('/login'));
-            }
-        } else {
-            $session->setFlashdata('msg', 'Email not Found');
+        $client = \Config\Services::curlrequest();
+            $r = $client->post('https://stockis.vercel.app/api/admin/login', [
+                'json' => [
+                    "email" => $email,
+                    "password" => $password,
+                ],
+            ]);
+            // Get the status code
+            $statusCode = $r->getStatusCode();
+            $body = $r->getBody();
+            $result = json_decode($body, true);
+            
+            $bolehhh = $statusCode == 202;
+            
+            // Output the response body and status code
+        if ($bolehhh) {
+            $ses_data = [
+                'user_id' => $result['data']['user_id'],
+                'user_name' => $result['data']['user_name'],
+                'user_email' => $result['data']['user_email'],
+                'logged_in' => true
+            ];
+            $session->set($ses_data);
+            return redirect()->to(base_url('/'));
+        }
+        else {
+            $session->setFlashdata('msg', "failed login");
             return redirect()->to(base_url('/login'));
         }
     }
@@ -52,7 +57,13 @@ class Login extends BaseController
     public function logout()
     {
         $session = session();
-        $session->destroy();
-        return redirect()->to(base_url('/login'));
+        if(session()->get('logged_in')){
+            $session->destroy();
+            return redirect()->to(base_url('/login'));
+        }
+        else {
+            $session->setFlashdata('msg', 'Please login first');
+            return redirect()->to(base_url('/login'));
+        }
     }
 }
